Rename TestForm.js to test-form.js as it is not a React component
